Returning to Basics
This commit removes empty files (that were preventing compilation) and
temporarily removes all non-core files. While the current commit still
doesn't compile (due to some non-implemented references) it is
significantly closer to where I want to be right now. I'd like to go
back to the very core of the application, building the GameMode,
GameState, GameInstance, GameSession, PlayerState, PlayerController, and
LocalPlayer classes first before even touching "Mahjong" code. It is
important to restart here to ensure the structure of my application is
correct as well as learn the necessary steps to creating an application
with networking features.

- Removed most empty files.
- Added GameUserSettings derived class.
- Updated the game modes to somewhat reflect their goals.
- Added very bare GameXXX and PlayerXXX classes.