Miscellaneous Commits
This commit includes various code refactoring and changes.

- Added "MahjongGame/Private" to private include paths.
- Moved EMap enum out of FMahjongMainMenu and into the global namespace
in its own file and renamed it to EMahjongMap. This way any class that
handles maps and chunks (for map streaming, paks, etc...) will have
convenient access to the enum. The map value given in the enum will act
as the chunk id unless this becomes an issue.
- Added MahjongGameDelegates.h that handles paks, save games, web
servers, and layer chunks.
- Added InitializeMahjongGameDelegates() function call to
MahjongGame.cpp (in StartupModule()).
- Added AMahjongPlayerStart to properly position players around the
table at the start of the game (and when they are killed off and respawn
between rounds where their winds change).
- Overrode ChoosePlayerStart implementation in MahjongGameMode to only
spawn players in a PlayerStart that corresponds to their wind.
- Uncommented FMahjongStyle::Initialize() in
MahjongPlayerController_Menu.cpp.
- Changed EMahjongPlayerWind to a UENUM to allow assigning its values in
the editor (since UENUM parsing in UE4 does not allow non-explicit
values - meaning I can't assign an enum value to another). This means I
needed to add a utility method to check whether a EMahjongPlayerWind
value matches a EMahjongTileId value since the winds will not have
matching values. This works out fine since the two enums are no longer
bound to each other.
- Removed commented UMahjongDeck code from MahjongGameMode_Japanese.h
since the class is now called UMahjongTilePool, and it doesn't belong in
the GameMode to begin with (it belongs in GameState).